+++ attributes/config.rb
@@ -7,6 +7,7 @@ default['stackstorm']['config']['api'] = {
   host: '0.0.0.0',
   port: 9101,
   logging: '/etc/st2api/logging.conf',
+  serve_webui_files: true
 }
 
 # Auth listener.
@@ -17,7 +18,8 @@ default['stackstorm']['config']['auth'] = {
   enable: false,
   use_ssl: false,
   mode: 'proxy',
-  logging: '/etc/st2auth/logging.conf'
+  logging: '/etc/st2auth/logging.conf',
+  api_url: %Q(#{node['stackstorm']['api_url']})
 }
 
 default['stackstorm']['config']['st2_webhook_sensor'] = {
@@ -33,16 +35,20 @@ default['stackstorm']['config']['generic_webhook_sensor'] = {
 }
 # ---->
 
+# St2 API URL
 default['stackstorm']['config']['sensorcontainer'] = {
-  actionexecution_base_url: "http://#{"#{node['stackstorm']['config']['api'][:host]}:#{node['stackstorm']['config']['api'][:port]}"}/actionexecutions",
-  logging: '/etc/st2reactor/logging.sensorcontainer.conf',
+  actionexecution_base_url: %Q(#{node['stackstorm']['api_url']}/actionexecutions/),
+  logging: '/etc/st2reactor/logging.sensorcontainer.conf'
 }
 
 default['stackstorm']['config']['rulesengine'][:logging] = '/etc/st2reactor/logging.rulesengine.conf'
 default['stackstorm']['config']['actionrunner'][:logging] = '/etc/st2actions/logging.conf'
-default['stackstorm']['config']['history'][:logging] = '/etc/st2actions/logging.history.conf'
+default['stackstorm']['config']['resultstracker'][:logging] = '/etc/st2actions/logging.resultstracker.conf'
 
-default['stackstorm']['config']['log'][:excludes] = 'requests'
+default['stackstorm']['config']['log'] = {
+  excludes: 'requests,paramiko',
+  redirect_stderr: false
+}
 default['stackstorm']['config']['system'][:base_path] = node['stackstorm']['home']
 
 default['stackstorm']['config']['ssh_runner'][:remote_dir] = '/tmp'
@@ -50,7 +56,8 @@ default['stackstorm']['config']['ssh_runner'][:remote_dir] = '/tmp'
 default['stackstorm']['config']['syslog'] = {
   host: 'localhost',
   port: 514,
-  facility: 'local7'
+  facility: 'local7',
+  protocol: 'udp'
 }
 
 
@@ -58,6 +65,5 @@ default['stackstorm']['config']['syslog'] = {
 default['stackstorm']['config']['messaging'][:url] = 'amqp://guest:guest@localhost:5672/'
 
 default['stackstorm']['config']['action_sensor'] = {
-  triggers_base_url: "http://localhost:9101/triggertypes/",
-  webhook_sensor_base_url: "http://localhost:6000/webhooks/st2/"
+  triggers_base_url: %Q(#{node['stackstorm']['api_url']}/triggertypes/)
 }