+++ attributes/mistral.rb
@@ -1,1 +1,0 @@
-override['openstack-mistral']['etc_dir'] = '/etc/mistral'