+++ recipes/_python.rb
@@ -1,9 +1,7 @@
 include_recipe "python::#{node['python']['install_method']}"
 
-if platform_family?("rhel", "fedora")
-  # st2* packages depend on python-pip from epel, so we work it around.
-  # Though packages should not strictly depened like this!
-  include_recipe 'yum-epel::default'
+if platform_family?('rhel', 'fedora')
+  include_recipe('yum-epel') if platform_family?('rhel')
 
   package('python-pip') do
     action :install