+++ recipes/install_repo.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: stackstorm
-# Recipe:: install_stackstorm
+# Recipe:: install_repo
 #
 # Copyright (C) 2015 StackStorm (info@stackstorm.com)
 #
@@ -12,20 +12,30 @@ include_recipe 'stackstorm::_initial'
 node.override['stackstorm']['bin_dir'] = '/usr/bin'
 node.override['stackstorm']['python_binary'] = node['python']['binary']
 
-cache_dir = "#{node['stackstorm']['home']}/package_cache"
-
-directory ":create package cache directory #{cache_dir}"  do
-  path cache_dir
-  owner 'root'
-  group 'root'
-  mode '0755'
-  action :create
+# Add StackStorm repository
+case node['platform_family']
+when 'debian'
+  apt_repository "downloads.stackstorm.net" do
+    uri node['stackstorm']['install_repo']['base_url']
+    distribution "#{node['lsb']['codename']}_#{node['stackstorm']['install_repo']['suite']}"
+    key "#{node['stackstorm']['install_repo']['base_url']}/pubkey.gpg"
+    components ['main']
+  end
+when 'fedora', 'rhel'
+  yum_repository 'downloads.stackstorm.net' do
+    description "StackStorm offical package repository"
+    baseurl node['stackstorm']['install_repo']['base_url']
+    gpgcheck false
+    # gpgkey "#{node['stackstorm']['install_repo']['base_url']}/pubkey.gpg"
+    action :create
+  end
 end
 
-stackstorm_fetch ':install stackstorm packages from St2 repository' do
-  path cache_dir
-  packages node['stackstorm']['install_stackstorm']['packages']
-  action :install
+# Install packages
+node['stackstorm']['install_repo']['packages'].each {|p| package p}
+
+package 'st2debug' do
+  only_if { node['stackstorm']['install_repo']['debug_package'] == true }
 end
 
 remote_file "#{node['stackstorm']['etc_dir']}/st2-requirements.txt" do