+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.2
+
+ * Native repository layout support
+ * Various fixes, tested on Fedora 20, Centos 70 and Ubuntu 14.04
+
 # 0.1.0
 
 Initial release of stackstorm cookbook